Reorganize types.py files
